Student scheduling
First attempt at getting students able to add classes to their schedule. Wanting the schedule stored on the database and associated with the user that added them so it can be brought up on the next log in. Added classes should display on the current schedule page where a student can then drop a class.